fix: set tag type to `uint8_t` (#437)
## Description

This PR updates the internal representation of cell tagging flags by
making `samurai::CellFlag` use an 8-bit underlying type and adjusting
call sites to cast to `std::uint8_t` when writing/reading tag bitmasks.

**Changes:**

- Change `CellFlag` to `enum class CellFlag : std::uint8_t`.
- Update tagging code paths to use
`static_cast<std::uint8_t>(CellFlag::...)` instead of
`static_cast<int>(...)`.

## Code of Conduct
By submitting this PR, you agree to follow our [Code of
Conduct](https://github.com/hpc-maths/samurai/blob/master/docs/CODE_OF_CONDUCT.md)
- [x] I agree to follow this project's Code of Conduct

+++ include/samurai/algorithm/utils.hpp
@@ -62,14 +62,14 @@ namespace samurai
             static_nested_loop<1, -s, s + 1>(
                 [&](const auto& stencil)
                 {
-                    tag(level, i + stencil[0]) |= static_cast<int>(CellFlag::keep);
+                    tag(level, i + stencil[0]) |= static_cast<std::uint8_t>(CellFlag::keep);
                 });
         }
 
         template <class T, class Flag, int s>
         SAMURAI_INLINE void operator()(Dim<1>, T& tag, const Flag& flag, std::integral_constant<int, s>) const
         {
-            auto mask = (tag(level, i) & static_cast<int>(flag));
+            auto mask = (tag(level, i) & static_cast<std::uint8_t>(flag));
 
             apply_on_masked(mask,
                             [&](auto imask)
@@ -77,7 +77,7 @@ namespace samurai
                                 static_nested_loop<1, -s, s + 1>(
                                     [&](const auto& stencil)
                                     {
-                                        tag(level, i + stencil[0])(imask) |= static_cast<int>(CellFlag::keep);
+                                        tag(level, i + stencil[0])(imask) |= static_cast<std::uint8_t>(CellFlag::keep);
                                     });
                             });
         }
@@ -88,14 +88,14 @@ namespace samurai
             static_nested_loop<2, -s, s + 1>(
                 [&](const auto& stencil)
                 {
-                    tag(level, i + stencil[0], j + stencil[1]) |= static_cast<int>(CellFlag::keep);
+                    tag(level, i + stencil[0], j + stencil[1]) |= static_cast<std::uint8_t>(CellFlag::keep);
                 });
         }
 
         template <class T, class Flag, int s>
         SAMURAI_INLINE void operator()(Dim<2>, T& tag, const Flag& flag, std::integral_constant<int, s>) const
         {
-            auto mask = (tag(level, i, j) & static_cast<int>(flag));
+            auto mask = (tag(level, i, j) & static_cast<std::uint8_t>(flag));
 
             apply_on_masked(mask,
                             [&](auto imask)
@@ -103,7 +103,7 @@ namespace samurai
                                 static_nested_loop<2, -s, s + 1>(
                                     [&](const auto& stencil)
                                     {
-                                        tag(level, i + stencil[0], j + stencil[1])(imask) |= static_cast<int>(CellFlag::keep);
+                                        tag(level, i + stencil[0], j + stencil[1])(imask) |= static_cast<std::uint8_t>(CellFlag::keep);
                                     });
                             });
         }
@@ -114,24 +114,25 @@ namespace samurai
             static_nested_loop<3, -s, s + 1>(
                 [&](const auto& stencil)
                 {
-                    tag(level, i + stencil[0], j + stencil[1], k + stencil[2]) |= static_cast<int>(CellFlag::keep);
+                    tag(level, i + stencil[0], j + stencil[1], k + stencil[2]) |= static_cast<std::uint8_t>(CellFlag::keep);
                 });
         }
 
         template <class T, class Flag, int s>
         SAMURAI_INLINE void operator()(Dim<3>, T& tag, const Flag& flag, std::integral_constant<int, s>) const
         {
-            auto mask = (tag(level, i, j, k) & static_cast<int>(flag));
+            auto mask = (tag(level, i, j, k) & static_cast<std::uint8_t>(flag));
 
-            apply_on_masked(mask,
-                            [&](auto imask)
-                            {
-                                static_nested_loop<3, -s, s + 1>(
-                                    [&](const auto& stencil)
-                                    {
-                                        tag(level, i + stencil[0], j + stencil[1], k + stencil[2])(imask) |= static_cast<int>(CellFlag::keep);
-                                    });
-                            });
+            apply_on_masked(
+                mask,
+                [&](auto imask)
+                {
+                    static_nested_loop<3, -s, s + 1>(
+                        [&](const auto& stencil)
+                        {
+                            tag(level, i + stencil[0], j + stencil[1], k + stencil[2])(imask) |= static_cast<std::uint8_t>(CellFlag::keep);
+                        });
+                });
         }
     };
 
@@ -161,58 +162,58 @@ namespace samurai
         template <class T>
         SAMURAI_INLINE void operator()(Dim<1>, T& tag) const
         {
-            auto mask = (tag(level + 1, 2 * i) & static_cast<int>(CellFlag::keep))
-                      | (tag(level + 1, 2 * i + 1) & static_cast<int>(CellFlag::keep));
+            auto mask = (tag(level + 1, 2 * i) & static_cast<std::uint8_t>(CellFlag::keep))
+                      | (tag(level + 1, 2 * i + 1) & static_cast<std::uint8_t>(CellFlag::keep));
 
             apply_on_masked(mask,
                             [&](auto imask)
                             {
-                                tag(level + 1, 2 * i)(imask) |= static_cast<int>(CellFlag::keep);
-                                tag(level + 1, 2 * i + 1)(imask) |= static_cast<int>(CellFlag::keep);
+                                tag(level + 1, 2 * i)(imask) |= static_cast<std::uint8_t>(CellFlag::keep);
+                                tag(level + 1, 2 * i + 1)(imask) |= static_cast<std::uint8_t>(CellFlag::keep);
                             });
         }
 
         template <class T>
         SAMURAI_INLINE void operator()(Dim<2>, T& tag) const
         {
-            auto mask = (tag(level + 1, 2 * i, 2 * j) & static_cast<int>(CellFlag::keep))
-                      | (tag(level + 1, 2 * i + 1, 2 * j) & static_cast<int>(CellFlag::keep))
-                      | (tag(level + 1, 2 * i, 2 * j + 1) & static_cast<int>(CellFlag::keep))
-                      | (tag(level + 1, 2 * i + 1, 2 * j + 1) & static_cast<int>(CellFlag::keep));
+            auto mask = (tag(level + 1, 2 * i, 2 * j) & static_cast<std::uint8_t>(CellFlag::keep))
+                      | (tag(level + 1, 2 * i + 1, 2 * j) & static_cast<std::uint8_t>(CellFlag::keep))
+                      | (tag(level + 1, 2 * i, 2 * j + 1) & static_cast<std::uint8_t>(CellFlag::keep))
+                      | (tag(level + 1, 2 * i + 1, 2 * j + 1) & static_cast<std::uint8_t>(CellFlag::keep));
 
             apply_on_masked(mask,
                             [&](auto imask)
                             {
-                                tag(level + 1, 2 * i, 2 * j)(imask) |= static_cast<int>(CellFlag::keep);
-                                tag(level + 1, 2 * i + 1, 2 * j)(imask) |= static_cast<int>(CellFlag::keep);
-                                tag(level + 1, 2 * i, 2 * j + 1)(imask) |= static_cast<int>(CellFlag::keep);
-                                tag(level + 1, 2 * i + 1, 2 * j + 1)(imask) |= static_cast<int>(CellFlag::keep);
+                                tag(level + 1, 2 * i, 2 * j)(imask) |= static_cast<std::uint8_t>(CellFlag::keep);
+                                tag(level + 1, 2 * i + 1, 2 * j)(imask) |= static_cast<std::uint8_t>(CellFlag::keep);
+                                tag(level + 1, 2 * i, 2 * j + 1)(imask) |= static_cast<std::uint8_t>(CellFlag::keep);
+                                tag(level + 1, 2 * i + 1, 2 * j + 1)(imask) |= static_cast<std::uint8_t>(CellFlag::keep);
                             });
         }
 
         template <class T>
         SAMURAI_INLINE void operator()(Dim<3>, T& tag) const
         {
-            auto mask = (tag(level + 1, 2 * i, 2 * j, 2 * k) & static_cast<int>(CellFlag::keep))
-                      | (tag(level + 1, 2 * i + 1, 2 * j, 2 * k) & static_cast<int>(CellFlag::keep))
-                      | (tag(level + 1, 2 * i, 2 * j + 1, 2 * k) & static_cast<int>(CellFlag::keep))
-                      | (tag(level + 1, 2 * i + 1, 2 * j + 1, 2 * k) & static_cast<int>(CellFlag::keep))
-                      | (tag(level + 1, 2 * i, 2 * j, 2 * k + 1) & static_cast<int>(CellFlag::keep))
-                      | (tag(level + 1, 2 * i + 1, 2 * j, 2 * k + 1) & static_cast<int>(CellFlag::keep))
-                      | (tag(level + 1, 2 * i, 2 * j + 1, 2 * k + 1) & static_cast<int>(CellFlag::keep))
-                      | (tag(level + 1, 2 * i + 1, 2 * j + 1, 2 * k + 1) & static_cast<int>(CellFlag::keep));
+            auto mask = (tag(level + 1, 2 * i, 2 * j, 2 * k) & static_cast<std::uint8_t>(CellFlag::keep))
+                      | (tag(level + 1, 2 * i + 1, 2 * j, 2 * k) & static_cast<std::uint8_t>(CellFlag::keep))
+                      | (tag(level + 1, 2 * i, 2 * j + 1, 2 * k) & static_cast<std::uint8_t>(CellFlag::keep))
+                      | (tag(level + 1, 2 * i + 1, 2 * j + 1, 2 * k) & static_cast<std::uint8_t>(CellFlag::keep))
+                      | (tag(level + 1, 2 * i, 2 * j, 2 * k + 1) & static_cast<std::uint8_t>(CellFlag::keep))
+                      | (tag(level + 1, 2 * i + 1, 2 * j, 2 * k + 1) & static_cast<std::uint8_t>(CellFlag::keep))
+                      | (tag(level + 1, 2 * i, 2 * j + 1, 2 * k + 1) & static_cast<std::uint8_t>(CellFlag::keep))
+                      | (tag(level + 1, 2 * i + 1, 2 * j + 1, 2 * k + 1) & static_cast<std::uint8_t>(CellFlag::keep));
 
             apply_on_masked(mask,
                             [&](auto imask)
                             {
-                                tag(level + 1, 2 * i, 2 * j, 2 * k)(imask) |= static_cast<int>(CellFlag::keep);
-                                tag(level + 1, 2 * i + 1, 2 * j, 2 * k)(imask) |= static_cast<int>(CellFlag::keep);
-                                tag(level + 1, 2 * i, 2 * j + 1, 2 * k)(imask) |= static_cast<int>(CellFlag::keep);
-                                tag(level + 1, 2 * i + 1, 2 * j + 1, 2 * k)(imask) |= static_cast<int>(CellFlag::keep);
-                                tag(level + 1, 2 * i, 2 * j, 2 * k + 1)(imask) |= static_cast<int>(CellFlag::keep);
-                                tag(level + 1, 2 * i + 1, 2 * j, 2 * k + 1)(imask) |= static_cast<int>(CellFlag::keep);
-                                tag(level + 1, 2 * i, 2 * j + 1, 2 * k + 1)(imask) |= static_cast<int>(CellFlag::keep);
-                                tag(level + 1, 2 * i + 1, 2 * j + 1, 2 * k + 1)(imask) |= static_cast<int>(CellFlag::keep);
+                                tag(level + 1, 2 * i, 2 * j, 2 * k)(imask) |= static_cast<std::uint8_t>(CellFlag::keep);
+                                tag(level + 1, 2 * i + 1, 2 * j, 2 * k)(imask) |= static_cast<std::uint8_t>(CellFlag::keep);
+                                tag(level + 1, 2 * i, 2 * j + 1, 2 * k)(imask) |= static_cast<std::uint8_t>(CellFlag::keep);
+                                tag(level + 1, 2 * i + 1, 2 * j + 1, 2 * k)(imask) |= static_cast<std::uint8_t>(CellFlag::keep);
+                                tag(level + 1, 2 * i, 2 * j, 2 * k + 1)(imask) |= static_cast<std::uint8_t>(CellFlag::keep);
+                                tag(level + 1, 2 * i + 1, 2 * j, 2 * k + 1)(imask) |= static_cast<std::uint8_t>(CellFlag::keep);
+                                tag(level + 1, 2 * i, 2 * j + 1, 2 * k + 1)(imask) |= static_cast<std::uint8_t>(CellFlag::keep);
+                                tag(level + 1, 2 * i + 1, 2 * j + 1, 2 * k + 1)(imask) |= static_cast<std::uint8_t>(CellFlag::keep);
                             });
         }
     };
